[update]拡大した状態でも画面移動ができるように(問題点有)
問題点：縮小した際に壁紙外が見えてしまう

+++ プロトタイプ/Camera.h
@@ -26,7 +26,7 @@ public:
 	int m_zoomCnt;
 	const double ZOOM_LIMIT_UP = 719.5;		//カメラの最大縮小値 1280x720が丁度表示される
 	const double ZOOM_LIMIT_DOWN = 300.0;	//カメラの最小拡大値
-	const double ZOOM_SCALE = 100.0;		//拡大縮小の移動値
+	const float ZOOM_SCALE = 100.0f;		//拡大縮小の移動値
 	int m_wheeled;							//過去のホイールの値の格納
 
 private:
@@ -46,8 +46,8 @@ private:
 
 	void CameraScale();
 	void CameraMove();
-	void CameraMoveX(double _moveAmount);
-	void CameraMoveY(double _moveAmount);
+	void CameraMoveX(float _moveAmount);
+	void CameraMoveY(float _moveAmount);
 	void CameraDrawControl();
 };
